+++ fe/src/App.tsx
@@ -12,7 +12,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Home from './pages/Home';
 import About from './pages/About';
 import Login from './pages/Login';
-import ProductPage from './pages/ProductPage';
+import ProductCreate from './pages/ProductCreate';
 import Error from './pages/Error';
 import Default from './components/layouts/Default';
 import ProductDetails from './pages/ProductDetails';
@@ -28,7 +28,7 @@ function App(): JSX.Element {
             <Route exact path="/" component={Home} />
             <Route path="/about" component={About} />
             <Route path="/Login" component={Login} />
-            <Route path="/Product/create" component={ProductPage} />
+            <Route path="/Product/create" component={ProductCreate} />
             <Route path="/product/:id" component={ProductDetails} />
             <Route component={Error} />
           </Switch>
