--- fe/src/pages/ProductPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProductPage';
